Backport XSA-471 patchset from Xen 4.17
This change backports the patchset for XSA-471 from Xen 4.17 to 4.13.

Patches 1002 to 1020 are backported from Xen 4.17 for some of the
XSA-471 patches to be meaningful.

2 patches from the original 4.17 patchset have not been backported.
These are xsa471-4.17-06.patch (x86/cpu-policy: Simplify logic in
guest_common_default_feature_adjustments()) and xsa471-4.17-15.patch
(x86/cpu-policy: Rearrange guest_common_*_feature_adjustments()) because
guest_common_*_feature_adjustments() are not implemented in 4.13 and
this drags too many changes to be backported.

Please see backport notes added to each patch file for details.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>