Disable CONFIG_REQUIRE_NX
At the moment the build flag CONFIG_REQUIRE_NX does nothing other than
preventing very old systems and systems that have the NX (Never eXecute)
set as "hidden" in the BIOS to not boot at all. But, even if NX is
hidden in the BIOS via 'MSR_IA32_MISC_ENABLE[34] = 1' on Intel or
'EFER.NXE = 0' on AMD, XEN is able to cope with this. XEN will first
probe the feature (via CPUID.80000001H:EDX[20]) if its 0, if will try to
set MSR_IA32_MISC_ENABLE[34] to 0, and will re-probe CPUID.80000001H,
effectively forcing NX to "enabled". So, there is no need for the
CONFIG_REQUIRE_NX, and there is also no security implication other than
for systems that don't have NX at all (32bit systems from before 2003)!

This issue affects our installer as well as a fully deployed XCP-ng
instances. If a user accidentaly flips this bit in the BIOS on a
deployed XCP-ng the system will not boot anymore! The boot code of XEN
does print a warning ("This build of Xen requires NX support") and then
reboots the system. However, this warning is only shown for a split
second if visible at all (depending on the speed of the CPU). So, the
user just sees the system rebooting and XCP-ng not working.

Signed-off-by: Julian Vetter <julian.vetter@vates.tech>